Update Metro config to extends `watchFolders` & `resolver.extraNodeModules`, that it can load modules outside of the MobileApp tree
The app works if we only import non-react dependencies.

```typescript
import {getPlural} from '../monolith-shared'

// then use getPlural in the code
```

As soon as we import a hook from shared module, the red screen appear on
mobile with the error:

```
ExceptionsManager.js:44 Error: Invalid hook call. Hooks can only be called inside of the body of a function component. This could happen for one of the following reasons:
1. You might have mismatching versions of React and the renderer (such as React DOM)
2. You might be breaking the Rules of Hooks
3. You might have more than one copy of React in the same app
```

The most likely reason is that we have multiple versions of React resolved inside the mobile
project