Metro: make use of blacklistRE to teach metro not to import React from other package
We want to blacklist React from other packages to enter the Module
Resolution tree.

But if the dependency has `import React from 'react'`, it will no longer
being able to resolve React anymore.

That is where `extraNodeModules` path come in, here we teach Metro where
to find `react` when it is imported from file outside of the current tree.